Improvements by using structs
-Improved access to W65C22 by using a struct
-Implemented LCD instances using a struct

+++ W65C02/src/w65c22.h
@@ -12,7 +12,7 @@
 /// @brief Sets the mode (input/output) of a single port.
 /// @param port A port from zero to 15 where 0 to seven are A's ports and eight to 15 B's ports.
 /// @param mode The mode to set on this port
-void portMode(uint16_t port, uint8_t mode);
+void portMode(uint8_t port, uint8_t mode);
 
 /// @brief Sets the mode (input/output) of all ports on the chip.
 /// @param ports The mode of all 16 ports. The first bit is port zero and the last one port 15. Zero represents input and one output.
@@ -21,7 +21,7 @@ void portModes(uint16_t ports);
 /// @brief Sets the output value of a certain port.
 /// @param port A port from zero to 15 where 0 to seven are A's ports and eight to 15 B's ports.
 /// @param value The logical value to set.
-void  writePort(uint16_t port, uint8_t value);
+void writePort(uint8_t port, uint8_t value);
 
 /// @brief Sets the output values of all ports.
 /// @param ports The logical output value of all 16 ports. The first bit is port zero and the last one port 15.
@@ -30,10 +30,9 @@ void writePorts(uint16_t ports);
 /// @brief Gets the current input/output value of a single port.
 /// @param port A port from zero to 15 where 0 to seven are A's ports and eight to 15 B's ports.
 /// @return Zero means the port is unpowered and everything upward means it's powered. Minus one means the specified port is out of range.
-int8_t readPort(uint16_t port);
+int8_t readPort(uint8_t port);
 
 /// @brief Gets the current input/output values of all ports.
-/// @param addr The address of the ports register (use constants).
 /// @return The first bit is corresponding to port zero, the second bit to port one and so on. One means powered and zero unpowered.
 uint16_t readPorts(void);
